Usage

User guide

Installation

StateMachines.jl can be – once it is registered – installed with

pkg> add StateMachines

`The development version is installed with

pkg> add("https://github.com/pbayer/StateMachines.jl")

The package is then loaded with

julia> using StateMachines

System description

A system has different levels, which are used to hide complexity.

StateMachines.Id — Type.

Id

An identifier used to identify and to register system components.

source

StateMachines.System — Type.

An common supertype for Box, Block and StateMachine.

source

StateMachines.Box — Type.

Box(id::Id, clk::Clock)

A box is the outermost container of blocks and state machines. It has a clock, a composite state and one or more input/output channels. It cannot contain other boxes, but it can interact with other boxes.

Arguments, Fields

id::Id: an identifier string

clk::Clock: the system clock, this may be a clock shared with other boxes,

state::AbstractState:

gate::Dict{Id, Channel}: a dictionary of channels,

child::Dict{Id, System}: a dictionary of registered blocks or state machines.

source

StateMachines.Block — Type.

Block(id::Id, surr::System)

A block is a container of blocks and state machines. It has one or more input/output channels. It registers to a system Box or to an higher block. It has a composite state.

Arguments, Fields

id::Id: each block in a surrounding has to have an unique identifier,

surr::System: this links to the surrounding system Block or Box,

cstate::State: a composite state of the underlying system,

gate::Dict{Id, Channel}: a dictionary of channels,

child::Dict{Id, System}: a dictionary of registered blocks or state machines.

source

Behavior and StateMachine are used to describe the activities of a system

StateMachines.Behavior — Type.

Behavior

State machines implement different behaviors. Therefore we have this abstract type to differentiate between state machines and to correctly dispatch on them.

Example

using StateMachines, Simulate

struct B1 <: Behavior end              # define the behaviors
struct B2 <: Behavior end

S1 = Box("Box", 𝐶)                     # define the environment,
Blk1 = Block("Blk1", S1)               # maybe containing one or more blocks

SM1 = StateMachine("SM1", B1(), Blk1)  # create state machine variables with
SM2 = StateMachine("SM2", B2(), Blk1)  # specified behaviors and suroundings

function step!(A::StateMachine{B1}, ::Idle, σ::Load)  # transition functions
    [....]                                            # implement the behaviour
end
function step!(A::StateMachine{B1}, ::Busy, σ::Fail) = [....]
function step!(A::StateMachine{B1}, ::Failed, ::Repair) = [....]
[....]

while true                           # later at runtime ...
    event = put!(SM1.gate[“in”])     # an event is read from an input channel
    step!(SM1, event)                # and a state machine transition is called
end

source

StateMachines.StateMachine — Type.

StateMachine{M}(id::Id, m::M, surr::System) where {M <: Behavior}
StateMachine(id::Id, m::Behavior, surr::System)

A state machine operates a Behavior. It has at least one input/output channel. It registers to a Block or a System.

Arguments, Fields

id::Id: each process in a box or a block has to have an unique identifier,

m::M: a model identifier,

cstate::AbstractState: the composite state of the state machine,

surr::System: links to the surrounding Block or System,

gate::Dict{Id, Channel}: events and tokens flow through the gates, each gate has an unique identifier.

state::State: the internal state is used to operate the state machine,

var::Dict{Id, Any}: a dictionary of local variables, each with a unique Id.

child::Dict{Id, System}: a dictionary of registered sub machines.

source

State machines

States

Events

Transitions

StateMachines.step! — Function.

step!(A::StateMachine, q::AbstractState, σ::AbstractEvent)

transition function δ causing a state machine A in state q₁ at event σ to take on a new state q₂.

For all specified transitions Δ: Q × Σ → P(Q) of A a step!-function has to be implemented. For unspecified transitions a fallback step function is called and a warning is generated.

Arguments

A::StateMachine: a state machine

q::AbstractState: any state ∈ Q

σ::AbstractEvent: any discrete event ∈ Σ

source

Operation