Usage

User guide

Installation

StateMachines.jl can be – once it is registered – installed with

pkg> add StateMachines

`The development version is installed with

pkg> add("https://github.com/pbayer/StateMachines.jl")

The package is then loaded with

julia> using StateMachines

System description

A system has different levels, which are used to hide complexity.

StateMachines.Id — Type.

Id

An identifier used to identify and to register system components.

source

StateMachines.System — Type.

An abstract type used to identify system components

source

Behavior and StateMachine are used to describe the activities of a system

StateMachines.Behavior — Type.

Behavior

Behavior differentiates state machines and is needed to correctly dispatch on them.

Example

using StateMachines, Simulate

struct B1 <: Behavior end               # define the behaviors
struct B2 <: Behavior end

SM1 = StateMachine("SM1", B1())         # create state machine variables with
SM2 = StateMachine("SM2", B2())         # specified behaviors and suroundings

function step!(A::StateMachine{B1}, ::Idle, σ::Load)  # transition functions
    [....]                                            # implement the behaviour
end
function step!(A::StateMachine{B1}, ::Busy, σ::Fail) = [....]
function step!(A::StateMachine{B1}, ::Failed, ::Repair) = [....]
[....]

while true                             # later at runtime ...
    event = put!(SM1.gate[“in”])       # an event is read from an input channel
    step!(SM1, event)                  # and a state machine transition is called
end

source

StateMachines.StateMachine — Type.

StateMachine{B}(id::Id, b::B) where {B <: Behavior}
StateMachine(id::Id, b::Behavior)

A state machine implements a Behavior. It can register to a complex state in another state machine.

Arguments, Fields

id::Id: each process in a box or a block has to have an unique identifier,

m::M: a model identifier,

env::Union{AbstractState, Nothing}: links to the surrounding complex state,

gate::Dict{Id, Channel}: events and tokens flow through the gates, each gate has an unique identifier.

state::AbstractState: used to operate the state machine,

var::Dict{Id, Any}: a dictionary of local variables, each with a unique Id.

source

State machines

States

StateMachines.AbstractState — Type.

AbstractState

A supertype to different state types. A state stands for a characteristic set of activities of a state machine. A set of states describes the state space of a state machine. An event may cause a state machine to change its state.

State types can be shared between different state machines. But they may mean different things since they stand for different activities.

There are different subtypes of AbstractState:

SimpleState,

CompositeState and OrthogonalState,

SubmachineState and

PseudoState

source

StateMachines.SimpleState — Type.

A simple state has no internal vertices or transitions.

source

StateMachines.CompositeState — Type.

A composite State contains at least one region.

source

StateMachines.OrthogonalState — Type.

An orthogonal state is a composite state containing multiple regions.

source

StateMachines.SubmachineState — Type.

A submachine State refers to an entire state machine nested within.

source

StateMachines.PseudoState — Type.

PseudoState

A Pseudostate is an abstraction for different transient states. A state machine enters a pseudo state temporarily within a transition to the next stable state.

source

Events

Transitions

StateMachines.step! — Function.

step!(A::StateMachine, q::AbstractState, σ::AbstractEvent)

transition function δ causing a state machine A in state q₁ at event σ to take on a new state q₂.

For all specified transitions of A a step!(A, q, σ)-function has to be implemented. The fallback step function does nothing and is called for all other calls to step!(A, q, σ).

Arguments

A::StateMachine: a state machine

q::AbstractState: any state ∈ Q

σ::AbstractEvent: any discrete event ∈ Σ

source

Operation