Usage

User guide

Installation

StateMachines.jl can be – once it is registered – installed with

pkg> add StateMachines

`The development version is installed with

pkg> add("https://github.com/pbayer/StateMachines.jl")

The package is then loaded with

julia> using StateMachines

System description

A system has different levels, which are used to hide complexity.

StateMachines.System — Type.

An abstract type for Box, Block and Process.

source

StateMachines.Box — Type.

Box(id::Id, clk::Clock)

A box is the outermost container of blocks and processes. It has one or more input channels and one or more output channels. It cannot contain other boxes, but it can interact with other boxes.

Arguments, Fields

id::Id: an identifier string

clk::Clock: the system clock, this may be a clock shared with other boxes,

state::State:

gate::Dict{Id, Channel}: a dictionary of channels,

childs::Dict{Id, System}: a dictionary of registered blocks or processes.

source

StateMachines.Block — Type.

Block(id::Id, surr::System)

A block is a container of blocks and processes. It has one or more input and one or more output channels. It registers to a System.

Arguments, Fields

id::Id: each block in a surrounding has to have an unique identifier,

surr::System: this links to the surrounding Block or System,

state::State:

gate::Dict{Id, Channel}: a dictionary of channels,

childs::Dict{Id, System}: a dictionary of registered agents.

source

StateMachine and Process are used to describe the activities of a system:

StateMachines.StateMachine — Type.

StateMachine

An abstract type for StateMachines. It is used to declare concrete state machines in an implementation.

Example

using StateMachines, Simulate

struct SM1 <: StateMachine end       # first you declare some state machine types
struct SM2 <: StateMachine end       # you want to use

S1 = Box("Box", 𝐶)                   # then you build a system,
B1 = Block("B1", S1)                 # maybe containing one or more blocks

P1 = Process("P1", SM1(), B1)        # processes inside a block act as state machines
P2 = Process("P2", SM2(), B1)

function step!(A::Process{SM1}, ::Idle, σ::Load)  # transition functions define the
    [....]                                        # activities of a process, maybe
end                                               # using other functions
function step!(A::Process{SM1}, ::Busy, σ::Fail) = [....]
function step!(A::Process{SM1}, ::Failed, ::Repair) = [....]
[....]

while true                           # later at runtime
    event = put!(P1.gate[“in”])      # an event is read from an input channel
    step!(P1, P1.state, event)       # and a process transition is called
end                                  # depending on its state and an incoming event

source

StateMachines.Process — Type.

Process{SM}(id::Id, sm::SM, surr::System) where {SM <: StateMachine}

A process is a container for a state machine. It cannot contain other processes. It has one input and one output channel. It registers to a Block or a System.

Arguments, Fields

id::Id: each process in a box or a block has to have an unique identifier,

sm::SM: a state machine identifier

surr::System: links to the surrounding Block or System,

state::State: this is read and changed by the process's state machine,

gate::Dict{Id, Channel}: events and tokens flow through the gates, each gate has an unique identifier.

source

State machines

States

Events

Transitions

Operation